OCRmypdf_plugins/ocrmypdf-surya-main/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocrmypdf-surya"
version = "0.1.0"
description = "Surya OCR plugin for OCRmyPDF"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MPL-2.0" }
authors = [{ name = "img_2_pdf contributors" }]
dependencies = [
  "ocrmypdf>=16.0.0,<17.0.0",
  "surya-ocr>=0.17.0,<0.18.0",
  "pillow>=10.0.0,<11.0.0",
  "pypdfium2==4.30.0",
]

[tool.setuptools]
packages = ["ocrmypdf_surya"]
